XIT FLOW: Apply review feedback
- Extract the shared price format into format.ts (FlowRow + PriceCell)
- Keep unpriced rows last only in the Value sort so the Ticker sort
  preserves category grouping
- Replace the nested ternary in formatAmount with an if/else chain
- Drop the redundant String() around the text input draft, which is
  always a string now that the input is type=text
- Rename .hiddenText to .text per the class naming rule
